Complete rewrite. Interrupt-based radio reading, asynchronous light settings, simplified user configuration, intergrated libraries.

+++ OpenSourceLights/AA_LIGHT_SETUP.ino
@@ -6,7 +6,7 @@
 
     // EXPLANATION OF STATES
     // ------------------------------------------------------------------------------------------------------------------------------------------------>
-    // Each light can respond to 14 different conditions, or states. These are: 
+    // Each light can respond to multiple different conditions, or states. These are: 
     // - Channel 3 - Position 1    
     // - Channel 3 - Position 2
     // - Channel 3 - Position 3    (This is the middle position if using a 3-position switch)
@@ -25,7 +25,9 @@
     //
     // State Notes: 
     // - At the very least, you must plug in the Throttle channel. Steering and Channel 3 are optional
-    // - If you do not plug in a steering channel, then obviously you will never encounter the Right Turn & Left Turn States
+    // - If you do not plug in a steering channel, then obviously you will never encounter the Right, Left and No Turn states. If you don't want to use 
+    //   the steering wheel but have another 3 position switch on your transmitter, you can plug that channel into the steering input and use the three 
+    //   states (Left turn, Right turn, No turn) for the 3 switch positions.
     // - If you do not plug in a third channel, the program will only use the values specified in Position 1 as the default. In other words, it acts
     //   as if you have a one-position switch that is always in Position 1
     // - If your third channel is only a 2-position switch, the program will switch between Position 1 and Position 5
@@ -38,18 +40,19 @@
     // For each light, you need to specify what will happen in each state - this is called the setting. The possible settings are: 
     // - ON
     // - OFF
-    // - FADEOFF        * Fades off slowly
+    // - FADEOFF        * Fades off slowly, time set in AA_UserConfig.h (FadeOutTime)
+    // - FADEON         * Fades on slowly, time set in AA_UserConfig.h (FadeInTime)
     // - NA
-    // - BLINK            Will blink at the rate set in UserConfig.h (BlinkInterval)
-    // - SOFTBLINK        Will blink at the rate set in UserConfig.h (BlinkInterval), but the light will fade-in and fade-out. The timing of the fade can be adjusted, see UserConfig 
-    // - FASTBLINK        Will blink at the fast rate set in UserConfig.h
-    // - DIM            * Will dim to the level set in UserConfig.h
+    // - BLINK            Will blink at the rate set in AA_UserConfig.h (BlinkInterval)
+    // - SOFTBLINK        A blink that recreates the look of old incandescant light bulbs. The timing of this blink is not user adjustable. 
+    // - FASTBLINK        Will blink at the fast rate set in AA_UserConfig.h (FastBlinkInterval)
+    // - DIM            * Will dim to the level set in AA_UserConfig.h (DimLevel)
     // - XENON          * Turns a light on but with a special effect that looks like a xenon bulb turning on
-    // - BACKFIRE         Special effect that blinks a light randomly for a short period of time (only works under Decelerating state - use for tailpipe LEDs)
+    // - BACKFIRE         Special effect that blinks a light randomly for a short period of time (use this under the Decelerating state for tailpipe/muffler LEDs)
     //
     // Settings Notes: 
     // - For the positions determined by Channel 3, it is best to specify an explicit setting, in other words, you probably don't want any of them to be NA - 
-    //   they should be ON, OFF, whatever. 
+    //   they should be ON, OFF, whatever.
     // - The opposite is true for the other states - if you don't want anything to happen during the Forward state for example, set it to NA, not OFF. 
     //   Otherwise, your light will turn off when going forward. 
     // * Some settings require a special feature known as PWM. These are marked above with an asterisks (*). Not all of the lights on the board are capable of implementing PWM, 
@@ -59,17 +62,16 @@
     // EXPLANATION OF SCHEMES
     // ------------------------------------------------------------------------------------------------------------------------------------------------>
     // A Scheme defines every setting for every light in every state. You may have as many Schemes as code space allows. Two are provided below to start, 
-    // if you want more, add them below, but remember to also update the NumSchemes variable in AA_UserConfig.h.     
+    // if you want more, add them below, but remember to also update the NumSchemes variable at the very top of AA_UserConfig.h
    
     // HOW TO SETUP YOUR LIGHTS
     // ------------------------------------------------------------------------------------------------------------------------------------------------>
     // Below you will see the lighting schemes. Each Scheme has a single row for each of the eight lights. The columns represent the states. The values
     // in the individual tables represent the settings for that light at that state. 
     // 
-    // OK, YOU'RE READY. TRY NOT TO MESS UP THE LAYOUT. JUST CHANGE THE SETTINGS. REMEMBER THE POSSIBLE VALUES: 
-    // ON, OFF, FADEOFF, BLINK, FASTBLINK, DIM, XENON, FADEOFF, NA
+    // OK, YOU'RE READY. TRY NOT TO MESS UP THE LAYOUT. JUST CHANGE THE SETTINGS. 
 
-    const PROGMEM uint16_t Schemes[NumSchemes][NumLights][NumStates] =
+    const PROGMEM uint8_t Schemes[NumSchemes][NumLights][NumStates] =
     {
         {
         //                                     IF CHANNEL 3 is only 3-position switch, values in Pos2 and Pos4 will be ignored (just use Pos1, Pos3, Pos5)
@@ -99,7 +101,6 @@
             {  OFF,          OFF,          OFF,          OFF,          OFF,          NA,           NA,           NA,           NA,           NA,           NA,           NA,           NA,           NA,           NA           }   // Light 8    -- 
         }
     };
-
